android_requirements.txt
# Core Requirements
kivy==2.2.1
kivymd==1.1.1
python-for-android==2023.9.1
pillow==10.1.0
requests==2.31.0
dnspython==2.4.2
ipwhois==1.2.0
speedtest-cli==2.1.3
ping3==4.0.4
pyOpenSSL==23.3.0
cython==3.0.5
buildozer==1.5.0

# Utility Libraries
uuid==1.30
matplotlib==3.7.2
seaborn==0.12.2
android==0.8.0
numpy==1.24.3
